+++ x/auth/client/cli/tx_sign.go
@@ -12,8 +12,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/tx"
-	kmultisig "github.com/cosmos/cosmos-sdk/crypto/keys/multisig"
-	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authclient "github.com/cosmos/cosmos-sdk/x/auth/client"
 )
@@ -435,9 +433,24 @@ func signTx(cmd *cobra.Command, clientCtx client.Context, txFactory tx.Factory, 
 			return err
 		}
 
-		fromRecord, err := clientCtx.Keyring.Key(fromName)
-		if err != nil {
-			return fmt.Errorf("error getting account from keybase: %w", err)
+		overwrite, _ := f.GetBool(flagOverwrite)
+		if multisig != "" {
+			multisigAddr, err := sdk.AccAddressFromBech32(multisig)
+			if err != nil {
+				// Bech32 decode error, maybe it's a name, we try to fetch from keyring
+				multisigAddr, _, _, err = client.GetFromFields(txFactory.Keybase(), multisig, clientCtx.GenerateOnly)
+				if err != nil {
+					return fmt.Errorf("error getting account from keybase: %w", err)
+				}
+			}
+			err = authclient.SignTxWithSignerAddress(
+				txF, clientCtx, multisigAddr, fromName, txBuilder, clientCtx.Offline, overwrite)
+			if err != nil {
+				return err
+			}
+			printSignatureOnly = true
+		} else {
+			err = authclient.SignTx(txF, clientCtx, clientCtx.GetFromName(), txBuilder, clientCtx.Offline, overwrite)
 		}
 		fromPubKey, err := fromRecord.GetPubKey()
 		if err != nil {